requirements.txt
python-telegram-bot
python-dotenv
aiofiles
httpx
